Move precice::utils::span to API as precice::span
