+++ src/sources.cmake
@@ -265,6 +265,7 @@ target_sources(preciceCore
     src/precice/impl/WatchPoint.hpp
     src/precice/impl/WriteDataContext.cpp
     src/precice/impl/WriteDataContext.hpp
+    src/precice/span.hpp
     src/precice/types.hpp
     src/query/Index.cpp
     src/query/Index.hpp
@@ -312,7 +313,6 @@ target_sources(preciceCore
     src/utils/fmtSTL.hpp
     src/utils/networking.cpp
     src/utils/networking.hpp
-    src/utils/span.hpp
     src/utils/span_tools.hpp
     src/utils/stacktrace.cpp
     src/utils/stacktrace.hpp
@@ -336,5 +336,6 @@ set_property(TARGET precice PROPERTY PUBLIC_HEADER
     ${PROJECT_BINARY_DIR}/src/precice/Version.h
     src/precice/SolverInterface.hpp
     src/precice/Tooling.hpp
+    src/precice/span.hpp
     src/precice/types.hpp
     )
