+++ src/utils/IntraComm.hpp
@@ -5,8 +5,8 @@
 #include "boost/range/irange.hpp"
 #include "com/SharedPointer.hpp"
 #include "logging/Logger.hpp"
+#include "precice/span.hpp"
 #include "precice/types.hpp"
-#include "utils/span.hpp"
 
 namespace precice {
 namespace logging {
